Track resource claims from jobs, both finished and evicted (#14)
* Use appropriate version of elasticsearch-python

* Track resource claims from evicted and finished jobs

* Use explicit scripted upsert for machine ads

This allows both machine and job updates to update the occupancy metric. Also,
the @timestamp field is now a list of every time the machine ad was updated.
This allows the doc to match multiple bins in a date histogram, making a kind
of hobo range indexing possible without upgrading to ES 7.

* Simplify resource->site mapping

Site names only need to be added in one place, and semantic names like
SLATE_US_UUTAH_NOTCHPEAK can simply be parsed.